fix#15 2

+++ Racing/Racing/Player.cs
@@ -9,7 +9,7 @@ namespace Racing
     class Player
     {
         public string name = "Player";
-        public int money = 1000;
+        public double money = 1000;
         public int rating = 0;
         public Player(int money, int rating,  string name)
         {
